rename-bff-to-mstd
