+++ mashup-savings-term-deposits/src/main/java/app/mstd/client/TimeDepositServiceClient.java
@@ -1,4 +1,5 @@
-package app.bff.client;
+
+package app.mstd.client;
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 
 import io.quarkus.oidc.token.propagation.AccessToken;
